+++ onprem/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.0.10"
+__version__ = "0.0.11"
 
 from .core import LLM